Refactor ASCE41Handlers and update tests
