+++ test/asce41_17-handler.spec.js
@@ -1,24 +1,10 @@
-/* global afterEach, beforeEach, describe, it */
+/* global describe, it */
 'use strict';
 
 
 const ASCE41_17Handler = require('../src/lib/asce41_17-handler'),
     expect = require('chai').expect,
     sinon = require('sinon');
-
-const _RESULT = {
-  data: [],
-  metadata: {}
-};
-
-const _FACTORY = {
-  destroy: () => {
-    // Nothing to do here
-  },
-  get: () => {
-    return Promise.resolve({data:[],metadata:{}});
-  }
-};
 
 
 describe('asce41_17-handler', () => {
@@ -37,52 +23,21 @@ describe('asce41_17-handler', () => {
       handler = ASCE41_17Handler();
       expect(handler.destroy).to.not.throw(Error);
     });
-  });
 
-  describe('checkParams', () => {
-    let handler;
+    it('sets the referenceDocument and instantiates the factory', () => {
+      let factory,
+          handler,
+          referenceDocument;
 
-    afterEach(() => {
-      handler.destroy();
-    });
-
-    beforeEach(() => {
-      handler = ASCE41_17Handler({factory: _FACTORY});
-    });
-
-    it('returns error if parameters are missing', (done) => {
-      handler.checkParams({}).then(() => {
-        return new Error('checkParams passed erroneously');
-      }).catch((err) => {
-        expect(err).to.be.an.instanceof(Error);
-        expect(err.status).to.equal(400);
-      }).then(done);
-    });
-  });
-
-  describe('get', () => {
-    let handler;
-
-    afterEach(() => {
-      handler.destroy();
-    });
-
-    beforeEach(() => {
-      handler = ASCE41_17Handler({factory: _FACTORY});
-    });
-
-    it('returns an object with data', (done) => {
-      sinon.stub(handler, 'checkParams').callsFake(() => {
-        return Promise.resolve({});
+      factory = sinon.spy();
+      referenceDocument = 'ASCE41-17';
+      handler = ASCE41_17Handler({
+        factory: factory,
+        referenceDocument: referenceDocument
       });
 
-      handler.get({}).then((params) => {
-        expect(params).to.deep.equal(_RESULT);
-      }).catch((err) => {
-        return err;
-      }).then(done);
-
-      handler.checkParams.restore();
+      expect(handler.referenceDocument).to.equal(referenceDocument);
+      expect(factory.called).to.be.true;
     });
   });
 });
